api/requirements.txt
flask
google-cloud-build==3.11.0
protobuf==4.25.3
